Add files via upload
The State of the Union Address is coming up, so for fun I calculated the sentiment of 30 yrs of speeches and plotted with economic growth. Obama's addresses seemed to closely reflect economic conditions in 2010 and 2011 following the crisis. By comparison, Biden's addresses during the pandemic seemed understated. What will the sentiment of Trump's speech look like?

The speeches are made at the beginning of the year, so president's were mostly speaking to economic conditions of the previous year. I standardized two sentiment scores from Afinn and Tblob and averaged them together for the sentiment index. Growth data is from the IMF's WEO database for the US.